121. Defining custom logging channel(Section 17: API Platform - Error Handling, Logging)
Now you have talked that channels are used to categorize the messages that are being locked.

Now let's see how we can create a custom Channel other than the default up Channel.

And other than the default Sinfonie like that doctrine or security channels.

So first create a monologue YAML file and under monologue.

Define a new channel.

Just other channels option and then other entry.

Maybe we'll call it Takhar up talkin confirmation.

Now let's use the console debug container and type log to see what are the services that can be categorized

as something that logs.

So now you can see under number 57 we have Manalo clogger token confirmation.

Now last time we have injected the logger interface in the constructor of our token confirmation service

and by default we got the logger for the app Jarno and this time we'll see how we can manually configure

our user conformations service and inject our new logger that will let us log using the user confirmation

channel.

So let's go to the services Yarm and specify our service manually by just using it's fully qualified

class name.

So we should specify the arguments of the service and we will use the variable name to only manually

specify one service that is being injected and everything else would be out of wire.

So we are out we are still using the outdoor wiring Sinfonie will automatically figure out what should

it be injected to that service.

The only thing we want to explicitly tell sinful what it should inject is their lager service.

So you should type monologue lager and the name forward the service was a monologue lager talkin confirmation.

So you can even copy and paste it here.

Now mind that I made a mistake here.

And the service should be actually injected by prepending its name with the sign

so you can see that there is this default monologue logger that will have that lock to that to the up

Channel.

So now after making the request you can see their mistake.

I've mentioned we are actually passing a string and the service should be prepared and we have to sign

so let's fix Stobbs.

And now we should be logging to our new channel.

OK so we've made a request.

It didn't work.

So now let's check our logs inside the var log dev data log and at the very bottom you can see that

there is a new channel called token confirmation with the debug level.

Now you can also see where this is useful.

We can use Stael command to open the def log file and filter all the error messages by specific channel.

So there are different places you can log your logs into but now let's just take a look at this example.

If you'd like to find only specific errors and in that case that would be errors from talkin consumation.

You can use grep together with Teyl which just opens a thousand lines with that and parameter and grep

is finding the string that you are searching for.

And you can this way.

Filter Oriol logs by their category.