122. Logging to separate files per channel(Section 17: API Platform - Error Handling, Logging)
Now at this point we were only using the basic stream blogger which looks to our file.

But before we move to something else let's also see how we can log their messages to different files

depending on their channel.

So it's pretty simple you just need to define another handler and it would be the handler of the same

type.

Because as I said the handlers are called from top to bottom.

So how do you define them in in what order you define them is important.

With exceptions that I will talk about later will specify the same path but this time instead of just

logging to General def dot logfile we will use a separate file for this channel which would be token

confirmation underscore their environment and to achieve this we just need to specify what channels

we would like to what channels should this Hendler log from.

And the only channel we would like this handler to be called is the token confirmation.

So now let's see what actually has changed.

So when you use Phylis for logging it might be useful to log different messages to different files because

they will grow exponentially very quickly.

So that might make a lot of sense to separate the locks into different files so we can see that we are

also logging the token confutation arrows to that log.

But other than that Additionally we are logging everything that is related to our token configuration

to our new separate file.

Now if you want to log it only to that particular file you will have to add this setting the same as

we are excluding that event channel.

In this dev log file now from the last request there is no message from the token configuration channel

because we have excluded it and now it is only exclusively logged to the token confirmation underscore

their current environment that log.