123. Production configuration for logger explained(Section 17: API Platform - Error Handling, Logging)
Now let's take a look at the production configuration for the monologue.

And you can notice that we it is much more complicated.

And the thing is that there are handlers that just put the locks blindly into a file or any other place

that you might specify.

But there are also handlers that filter them grouped them or do some additional processing before they

pass on the logs.

So one of those is that Fingers crossed.

So that's the first handler that you see defined.

And the way it works is that it will first wait for an error of a certain level to occur and in this

case it is level of error and only after any log with an error or level has been put to the log it will

only put the locks all the locks that were locked up until the time when an error occurred to the nested

handler.

So also notice that every handler that is nested inside another handler is not automatically called.

So as I said all these handlers work from top to bottom.

But let's take a look now.

But then nest that handler then the handle that has an exact name ness that is used as a handler inside

the main handler and also at the very bottom we can see that the forward handler is called deprecation

and it should look to stream but it is also on nested handler that is only being called Inside the deprecation

filter.

So DOS do not are not used automatically.

They only need to be triggered by the parent handler.

So going back to our fingers crossed.

So when we have this error or we have to with their severity of error.

So we have a log with severity of error or it is on me then being passed to the nested handler which

is itself writes to us stream to a file and it will log everything including the debug messages but

the fingers crossed is just triggering.