130. Installing PHPUnit and writing first basic unit test(Section 19: API Platform - Unit Testing (PHPUnit))
A good starting point with writing unit tests it's installing the library necessary to create those

tests run them and make assertions so let's install the BJP unit package for and then you can do that

as we do usually while using composer with their death floc and the name of the package is symphonic

slash ph B unit breach.

So wait until this installs that should be pretty quick.

And it says that to run the tests we can just run Pietsch be being slash BGB unit.

Now let's round ph P bin ph per unit for the first time.

This should create a new test folder for our project and it also might install some of its dependencies.

So as you see it's checking some of the dependencies and we have a message that no tests were executed.

And it makes sense we do not have any tests yet.

So there are tests for this Symphonie framework itself and all the other libraries.

But we are only running our own tests.

So additionally to saying that tests were executed or were successful we have the time it took and then

memory that was used.

So let's take a look at the composer Jason File right now and you can also see that apart from having

namespace definition for our project business logic the SLAC folder.

We also have a new namespace created there are slash tests and it is created for everything that leaves

inside that tests folder.

So this is the place where we'll be putting our test cases let's start straight away by creating a simple

test class.

Inside the tests folder and it should have the name space up tests.

Now every unit has to extend a certain class.

In our case it is the test case class from the BHB you need slash framework test case.

So that's the way one of the ways.

B You need to look for tests so if that class is extending that test case it is being treated as a class

with tests by the PH beyond.

So let's add the very simple test will test if the addition operator works.

So a single test would be a single method and it's a common practice to call this method test and then

something whatever is meaningful you'd like to verify.

Now let's other very simple assertions so we will check if two values are equal.

So the first argument is the value that you expect.

The second is the expression you'd like to run to see if it matches the expected value and the option

offered is the message.

So after running that as you can see that we have one test and one assertion.

So we said what our test is but assertion is a single check that is that leaves inside our test method

that actually verifies something.

So if you will modify the expected value to see you can see that we have their message 5:04 is expected

to equal to two plus plus free.

And there's also a message provided by pute beauty that it failed asserting that five matches expect

that six.

So you don't have to really write those custom messages for assertions.

You will come up with some defaults.

But if you like to make those metric messages more meaningful this makes sense to write an additional

message.

Now let's use another assertion.

Very simple one when we just check if a value is true.

So we have said the value variable to true.

So obviously this test will always work but we are not just checking if true is true.

We are seeing how we can.

What are the different kinds of assertions that we can make and one of the commonly used ones is the

answer equals.

And also true now let's create a simple array maybe just call it array variable we've key and Vario.

So there is another assertion we can use and it is US-CERT array Huskey.

This way we can check if a certain array has the given key.

So if you produce any kind of a race as a result of your functions or etc. You can check if they actually

return an array with the given key.

So that's also very useful.

Now there is a lot of those certain methods for a race but we can also verify if a Haskin.

So if this array would not have that in the test will just fail at this point.

So it makes sense to change those methods so first you verify if the array has the key.

And then you add another assertion with which you will check if the array value under the key is just

the same as you expect it to be.

So now also you can see that if we modify this array we have a nice output from that pH unit and it

says what value was expected and what it was.

What was actually in that expression.

So there is those two columns expected with minus sign and the actual plus sign.

So if you compare values exactly like you want to check if they match they're equal to each other you

would have this nice output which will tell you what was expected and what race or what was found actually.

And last session I'd like to show you in this lecture is the US-CERT count where you can check if the

count of objects inside an array for example is exactly how we expect it to be.