132. Mocks (stubs) Part 1 - Mocking dependencies, Entities(Section 19: API Platform - Unit Testing (PHPUnit))
In this lecture we will begin to talk about Mock's or stops.

So each class that we've created almost each service that we've created has some kind of dependencies

so you inject something from the constructor and our class is using those dependencies internally to

achieve something.

So for example we injected the token storage to our outer entity subscriber through the constructor

and then the token storage is being called to Fadge the current stock.

And then the current user.

So the unit tests can also check if this mettled is actually calling.

This gets doc and get user on the token storage and to achieve that we will actually use Mocs.

So we'll create a mocked or stabbed version of the token interface which is mind.

It is an interface.

So we are actually not stabbing or mocking the actual class but an interface but we can do it and this

will just test if the certain methods of that interface are called from within our method.

So we will run.

Actually the real method to get an integrated user but will pass some fake class do it.

That is based on the interface that it requires to see if it calls the right methods.

So lets start with how we create those marks.

So inside our test classes we have a method called Get Moq builder and the get mob builder accepts a

class name which is a fully qualified class name so you can use that Constanta that can called class

that is attached to all the classes and all that interfaces mph this will just returned the screen with

the name of the class and then for interfaces or abstract classes that might they cannot be instantiated.

So if you would like to create mark for interface or abstract class you will then call a chain method

call from the get model builder which itself returns a new instance that you can later do something

with.

But we just go get more for abstract abstract class and then we can safely pass this created mark to

the constructor of that outer entity subscriber.

Now to verify that certain medaled of up class or our dependency was actually called we can use expectations.

So on this newly created token Mark we have a method called expects inside it.

It accepts an argument which we can get from this once mettled from our class which one expect that

the feeling that we are declaring here should be called only once and exactly once.

There are other methods like that so we can use this one so there's never or this exactly where you

will specify how many times exactly you would want this method to be called.

And after that we use that metalled metal to provide the name of the method we expect to be called.

And this is get Dawkin.

And finally we have the will return when we can tell the smoke what it should return.

So we are not actually checking what it will return.

We can just set a value that that the call to the get talkin will actually return.

So we are just using it to HEFCE of scenarios like we have data that we might expect and we have two

or three use cases where different data is going through or is being returned.

So we can ride this mock Mock's or Stubb's to just test this case is what will happen there.

So now we are creating another mark because as you've seen that talkin storage interface itself returns

a talkin and then only the token has mettled get user.

So we have to also mark that talken interface class which is being returned from the get talkin mettled.

So now inside will return.

You'll see that from that dock and storage will return a mug of the talken interface which will have

a mettled get user the market mettled which is not really real metal.

It's something we created that will return a new instance of the user interface so that's what it is

returning.

In our case that we want to test we can have different cases and then try to return different objects

different values to test all the possible scenarios that this that thinks can take inside this method.

So we'll also try.

Or Usually you will also try to provide return return values that might cause errors so will you will

want to cover all the scenarios.