133.1 Mocks (stubs) Part 2 - Mocking dependencies, Entites(Section 19: API Platform - Unit Testing (PHPUnit))
Now last time we have created two mocks and explained how the work in general.

Now let's see what happens when we run this test.

So since we are not passing our marks to any class we haven't created our outer entity subscriber those

smogs actually create some expectations.

Those are basically assertions so we try to check if those methods are being called by by simply creating

the Mocs themselves.

They don't do anything.

We only need to pass them to the outer entity subscriber which itself accepts talkin storage interface.

But now since we created this Mock's we will pass our stops to this outer entity subscriber mind that

I am using Mock's and stops exchanging these words.

Thats basically the same.

And if you would like to call that get out integrated user We actually also have to provide an argument

because it accepts an event argument which is a type get response for controller result event.

You can see that on the bottom part of the screen that's the argument.

So we should also create a marks for this class sometimes Mock's can be useful if you do not really

want to create a whole complicated set of classes because sometimes all of these classes require themselves

some dependencies and now you can see that we have this case that get response for controller result

event constructor requires.

As I remember at least three parameters and we have passed zero.

So this could be an endless work when you try to mock all the possible or thats even more.

Its one two three four four parameters so that would be an overkill if we would like to pass all these

parameters or create real classes or mocs so we can use the sub disable original constructor and now

its just not passing anything.

So this class mind that when you create a mock this class will have all of its original Metellus.

So basically this just won't do anything.

Now we have another alert that the

that our subscriber expects that they're token that its parameter its constructor parameter has to be

of type specific interface type.

Its saying that in the console what it expects.

And our problem was that Ive just imported the wrong token storage interface so sorry for that.

That was just wrong wrong import inside the use statement.

Now we have another error and you can see on the bottom that our get out integrated user method is calling

get control controller result Mattox on the event and also get request on the event.

And weve created a mock that just doesnt have this method so its empty.

We only have to mock this metal's for it to run

and you can see that by commenting out those first two calls inside get out integrated user and returning.

So mind we are just temporarily modifying this method to see that our books actually work.

So we expect it gets documented to be called once and return book and interface and it did and then

it took an interface we expected the method to get user to be called once.