133.2 Mocks (stubs) Part 2 - Mocking dependencies, Entites(Section 19: API Platform - Unit Testing (PHPUnit))
So if you will now modify the talkin interface mock and instead of this one's called this never our

expectation is now that that gets you there would never be called and we called once so that this is

failing.

Now let's also see how we can expect that it'll be called exactly 2 times.

So looking at the code in the lower window you'll see that it is being called only once.

And our expectation says that we expect it to be called two times.

Exactly.

So that's how you can expect something to be called zero or one time or more times.

But we are calling it just once.

So let's leave our test.

Correct.

And this is passing as it is right now but we have to somehow solve the event that we are passing to

this method so we can also test if they get control and result and they get the request can be called

on our event class that we are passing to our test.