134.2 Event class Mock, extracting repeated mock factory code(Section 19: API Platform - Unit Testing (PHPUnit))
Now by changing the expectation to lever and running at this again U.S. got our initial test was correct.

So this is expected to be called once and it is being called once in the class so our test was correct

when we change it to never that this well started failing.

So let's change back to get control result met that expectation to be called once and we'll continue

with this in the next lecture.