135.1 Parameters for mocks(Section 19: API Platform - Unit Testing (PHPUnit))
So now let's go back to our get even more committed inside the test.

So the case we solved last time you start the event that is get passed through to get authenticated

user method calls that get control a result on that event.

So we have Moktar that mattered and provided a case where it returns it as it should be called once

and it will return a new instance of the block proposed.

So this mettled could continue and would be fine but there is also one morphemic that event has a GET

request mettled.

So to return that request from the event gets request we need to create a request mark so the request

Mark will also have mettled which is called Get method.

By the way.

So we need to mark all this objects for our expectations to work so we can check if everything is actually

called properly.

And later we'll use different cases where they get requests get that will not only return post it will

on also return get.

So we'll check if our or our cases work.

So we have our request Mark and the request Mark has method called Get method which returns a string

whether it is post method or guest medaled etc. so we can add another expectation to our event Mark

that it expects once the mettled

gets request to be called

and that it will return.

So this is basically not an expectation or we expectation is when we expect it to be called once.

But we are also saying that it will return a request.

MARK So you see that now that test is failing.

And this is because we have commented about this in the original code.

When we fix it it will work.

So now we have all of our expectations working mind that we still have an early read added to our get

out indicated user mettled not to go so so far we we are just testing the first three lines of this

matter.

So we have marked everything.