135.2 Parameters for mocks(Section 19: API Platform - Unit Testing (PHPUnit))
So that event get control a result is marked and is returning to the proper entity.

Now let's add that parameter to our Get event Mark which is creating our mark for the event.

And we can specify the parameter which would be what what is the name of the matter that we expect that

that event gets request get metalled will return.

Also we can specify what would be the controller results so what will be the entity returned from that

event get control the result.

As you remember.

This could be a blog post or image or any other entity.

So this model does get that indicated user will actually be called on every single of our entities when

they are persistent.

So we should check if it only works for the cases we wanted to.

So we see that we have an error and it is that too few arguments will pass to get event Mark.

So we are calling this mettled me but we modified its signature and haven't modified the arguments so

let's do the arguments that we had initially which is the post-Mandela.

And it should read our blog post and now you see that test is passing for this case which means whenever

there's a get out integrated user mettled inside out the outer entity subscriber will get will be called

we've post-marital.

And this would be for the blog post entity.

This will work.

So in the next lecture we will test different scenarios and different steps that could that the get

out indicate that user could get into.