136.1 Full test for AuthoredEntitySubscriber(Section 19: API Platform - Unit Testing (PHPUnit))
Now we are not yet testing.

The most important functionality of our outer entity subscriber inside the get out indicate user method

is shoot cold or set out or on the anti-Tea only if the entity implements particular interface or the

end.

The method is post so we have to create couple of scenarios.

First let's see what happens when we just copy it.

This chunk of code that creates new subscriber without creating a separate mocke for it.

So since then Mach was already used for this first case then coding again the methods on this mock caused

an error.

So the error set up it was called more than once when it shouldn't.

So let's fix it.

And by creating another set of marks which by which I mean calling the get and storage mark and get

even more Cagan and you can see that we have defined data for to get the method that event Mark will

return to get mettled that the request itself is to get met up.

And the entities block Post and our tests are still passing.

So that's not good.

It means we aren't checking the core functionality of this so we only got to the first line of this

get authenticated user when we check if the outer is being said from the token storage.

So how can we check whether the set out or is called on a particular entity or not.

So let's start by creating a mark on the block post-collapse itself.

So this time we are not building an abstract Mock from abstract class but from a concrete class which

is the block post.

We said that we want to mark the set Outer Limits.

And now on this entity mock we will we said that we require the set out are to be called once and now

after creating this entity maake we can parse it instead of actually creating a real instance of that

blog post.

We will pass this mark now.

Only now we can see if this is actually the code inside code get authenticated user is actually working

correctly.

So after running the test you see that we have one more assertion.

And that was the assertion for our mark of a blog post which for the first case that we had which was

the post dealt with and we've the instance of block post that it should call coal's set out or on the

entity.

Now if you will modify the method to get in the first case.

Now you see that test is failing so it knows that it shouldn't call that set out on the entity if it

is an instance of block post which into which implements the outer entity interface.

But the method is different so the method was get.

So this this this is working.

And we have this case tested.

So when there is a post-marital and a block post which actually implements that interface this is working

correctly.