136.2 Full test for AuthoredEntitySubscriber(Section 19: API Platform - Unit Testing (PHPUnit))
Now let's copy this part of code that creates the entity mocke and let's create a mockup builder for

a class that does not exist.

So you see that it brings.

It gives us an error on the right you can see that it says that we are trying to configure a method

set out or which cannot be configured because it does not exist because nonexisting class does not exist.

So we have to call set mettled first.

If we are trying to create a market for a non-existing class and we are creating an nonexisting class

because it does not implement the outer entity interface so we can test it for a different entity.

They set out or want to be cold.