136.3 Full test for AuthoredEntitySubscriber(Section 19: API Platform - Unit Testing (PHPUnit))
So now we have extracted the get entity mock to a separate method and instead of repeating this code

repeating this code all the time and creating either an instance of a blog post or and then deciding

if the set out or should be code or not.

You can see that it would produce a lot of code.

So we move that into a separate method which we can just call and get a new entity mark for our class

name that we want and we should specify if this should call their set out our method.

So we'll already know what we expect and we can use this use this get entity maake to create a lot of

different scenarios.

Now when the should call should call set out or risk drill will expect that the method should be called

once.

If it's false will expect that it should never be called.

Now let's go back to our test set out or call.

And now let's pass this arguments for our cases.

So the first case is that is the block post in.

And we expected to call the head out or because it is a post mettled.

And the second case is we'll be creating.

We'll be passing class does that just does not exist so we call it nonexisting.

And we expect that this set out or want to be called Let's also modified to get even Mach 2 to get our

new entity MMOG that we've created for the non existing class and we'll be going back to this test in

the next lecture.