137. Data providers in tests (using many data sets in 1 test case)(Section 19: API Platform - Unit Testing (PHPUnit))
You can see that now we are testing two scenarios.

One is when we have a POST method and entities that block posts we should call the set out and the entity.

And second case was a class that does not exist which is just different class that doesn't implement

out or entity interface and then set out or shouldn't be called.

So if we would like to create a Firth's scenario when we have a get metalled and that entity is block

post we should either paste the code that we already have and make dismantled unbearably long and imagine

adding even more scenarios that might be a little bit messy.

So there are better ways of testing different scenarios.

If you of course organize all your methods properly as we did before we're adding separate methods methods

that accept certain arguments for which we can easily create different Mock's that do different things

and just behave as we expect them to.

So there is a feeling inside Beej unique called our data provider.

So by using data provider for a single test you can just create like sets of different arguments and

expectations.

And by that I mean we could just set that this test should run for example three times and that it should

run for the set of arguments like block post with the post method nonexisting class with the get method

and block post we've get metalled so freak cases and that test set out or Colthurst will accept free

arguments.

So the first one would be the class name that will pass.

And that could be for example block post or something else.

So the second argument would be whether for this class it should call the set out or and there is also

a first parameter that we have which is the HTP metalled that we tested for.

So let's add this first argument to our tests that outer code and now instead of having this free identical

blocks where we define these different scenarios inside we will use data that is being passed through

the arguments.

So for the get entity Mark will use the class name the shoot code step out or and for the get event

MMOG to specify the mettled we'll use the mettled argument that we'll be passing the tests.

Outer code.

Now how we would be passing all those arguments if we will run the test right now you see that we have

an error.

It says that this test requires free arguments and we provided zero.

So the way we achieved that is we'll use an annotation called data provider and data provider has only

one setting.

It is the method that provides this sets of data and that's our provider set out or call method.

So the data provider should generally return an array and all the elements of this right should also

be an array as you see on the screen.

And those are the arguments that would be passed to the test said outhaul call as an argument in that

order you specified them.

So you see that the first elements of all our arrays is the class name.

The second is the Bodleian and deferred is the HTP method name.

And when you run this test again you'll see that it is passing so all of our scenarios are correct and

our test that test set out on call is very short and simple and it is being called ultimately for all

the free cases that we have provided through our data provider medaled which was provider set out to

call.