138.1 Finding potential errors using unit tests (null Token example)(Section 19: API Platform - Unit Testing (PHPUnit))
In the final lecture about uni there's we all see how we can find possible errors in our code by writing

tests.

So our Get talk and storage interface that we inject into the subscriber might also return.

And this comment says that if not out syndication information is provided that TACAN storage interface

can return either a token interface or no.

And we are not covering that case so we should write a test for it and then later update our code to

also handle this scenario.

So we are writing new test and we will call it test no talkin present for the cases where user might

not actually be authenticated mind that the user actually should be authenticated because we have security

constraints for for this particular for saving comments and users but let's just use it as an example.