138.2 Finding potential errors using unit tests (null Token example)(Section 19: API Platform - Unit Testing (PHPUnit))
So we are using all of our existing methods and this works not is that we werent creating a mock of

the entity class to sec to check if it will call the set out or because in this particular case we want

to get to that point because we want to check if that token storage gets taken my top will actually

return a token.

So lets add a new parameter to get took and storage mock the boolean parameter.

Call it hest Dokken.

And if that talken storage will have a token.

So if we set this parameter to true the mocked method to get token will actually create a token mark

so a mark for our token interface.

If this would be false it would return no.

So now inside our test no token present method.

Let's pass this argument and make it false.

So that would be the case when as you see on the right call to our member function get user on a null.

So this means there might be a case where the token storage get token actually returns on low end.

We are calling get user on no which is a critical error.

Impeach the.

So by writing this test we are able to modify our code and find potential issue with it.