138.3 Finding potential errors using unit tests (null Token example)(Section 19: API Platform - Unit Testing (PHPUnit))
So we should modify the way that we first fetch a token into the token variable and only if that token

is not known we continue.

But if the token is no we immediately return now modify also call that that is direct to the token storage

to use our token variable

Now if you run this test you can see that that expectation failed for mettled get use or when invoked

one time that the metalled was expected to be called once and it was called Zero times.

So this is because our get talkin storage maake actually has our token market that was created.

And if it is created at all this will have an expectation that will always run.

So if that has stock in is false we should modify it to expect mettled get user to be never called.