138.4 Finding potential errors using unit tests (null Token example)(Section 19: API Platform - Unit Testing (PHPUnit))
So after making this change we have also we have found an error inside our class and we also covered

every possible scenario.

So our code is covered hundred percent so it's another thing called code coverage.

So you should aim for the biggest percent of your code covered by tests and you can also create a difference

scenario as you see you have a comment class that will be sent using the post method in which case we

expect to set out to be cold.

And basically now for this single class we have everything tests that every line is covered by the tests.

So we we are checking for every possible scenario that can happen there we are even verifying that the

static configuration of this outer entity subscriber is correct and we have created seven tests and

have 31 assertions.

So if someone will modify the way this glass works the failing test will signal to him that he did something

wrong.

So that's the very useful thing of the unique test that you are always certain that logic that you have

implement that works exactly as you expected.

And the biggest advantage of that is that the others other members of your team that weren't writing

that code will be sure that whenever they change something.